Moved file back to original location
